initial pass to making code mirror/query input with a fixed width and height. created dev folder for code snippets not being used.
