+++ client/StyleSheets/_queryInput.scss
@@ -6,6 +6,7 @@
     grid-template-areas:
     "codeMirror"
     "buttons" ;
+    height: 70vh;
 
     #code-mirror {
         grid-area: codeMirror
@@ -19,9 +20,19 @@
 
 
     .react-codemirror2 {
+        max-height: 66.5vh;
         border-radius: 3px;
-        height: 100%;
         overflow-y: scroll;
+        overflow-x: scroll;
+
+        .CodeMirror-scroll{
+            // div inside of react-codemirror2 that dynamically changes based on code lines
+            height: 100%;
+            min-height: 72vh;
+            // move this max width somewhere else
+            max-width: 38vw;
+        }
     }
+
 
 }
